gogir-labs-be/requirements.txt
Django==4.2.29
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.3.0
django-ckeditor==6.7.3
drf-spectacular==0.26.5
django-filter==23.5
celery==5.3.4
redis==5.0.1
gunicorn==22.0.0
whitenoise==6.6.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
dj-database-url==2.1.0
django-redis==5.4.0
reportlab==4.0.7

